re-implement preferences dialog
