+++ src/lib/component/partial/index.js
@@ -51,7 +51,8 @@ export { default as NamespaceLabel } from "./NamespaceLabel";
 export { default as NamespaceSelector } from "./NamespaceSelector";
 export { default as NotFound } from "./NotFound";
 export { default as Pagination } from "./Pagination";
-export { default as QuickNav } from "./Drawer/QuickNav";
+export { default as Preferences } from "./Preferences";
+export { default as PreferencesDialog } from "./PreferencesDialog";
 export { default as RelatedEntitiesCard } from "./RelatedEntitiesCard";
 export { default as ResetAdornment } from "./ResetAdornment";
 export { default as ResourceDetails } from "./ResourceDetails";
